ext2_metadata: rename to clarify this is ext2 specific
Avoids confusion with other things like payload metadata.
