+++ src/update_engine/utils.h
@@ -313,17 +313,6 @@ class ScopedEintrSafeFdCloser {
   DISALLOW_COPY_AND_ASSIGN(ScopedEintrSafeFdCloser);
 };
 
-// Utility class to close a file system
-class ScopedExt2fsCloser {
- public:
-  explicit ScopedExt2fsCloser(ext2_filsys filsys) : filsys_(filsys) {}
-  ~ScopedExt2fsCloser() { ext2fs_close(filsys_); }
-
- private:
-  ext2_filsys filsys_;
-  DISALLOW_COPY_AND_ASSIGN(ScopedExt2fsCloser);
-};
-
 // Utility class to delete a file when it goes out of scope.
 class ScopedPathUnlinker {
  public:
